# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everruns API",
    "description": "API for managing AI agents, sessions, messages, and events",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://app.everruns.com/api",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/orgs": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "GET /v1/orgs - List organizations the current user belongs to",
        "operationId": "list_organizations",
        "responses": {
          "200": {
            "description": "List of organizations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_OrganizationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "POST /v1/orgs - Create a new organization",
        "operationId": "create_organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/v1/orgs/{org}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "GET /v1/orgs/:org - Get organization details",
        "operationId": "get_organization",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "Organization public ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "PATCH /v1/orgs/:org - Update organization",
        "operationId": "update_organization",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "Organization public ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Organization updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/v1/sessions/{session_id}/storage/keys": {
      "get": {
        "tags": [
          "session-storage"
        ],
        "summary": "GET /v1/sessions/{session_id}/storage/keys - List all key-value pairs",
        "operationId": "list_keys",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of key-value pairs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_KeyValueInfo"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/sessions/{session_id}/storage/secrets": {
      "get": {
        "tags": [
          "session-storage"
        ],
        "summary": "GET /v1/sessions/{session_id}/storage/secrets - List all secrets (names only)",
        "operationId": "list_secrets",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of secrets (names only, no values)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_SecretInfo"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganizationRequest": {
        "type": "object",
        "description": "Request to create a new organization",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The display name of the organization.",
            "example": "Acme Corp"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        }
      },
...
      "ListResponse_KeyValueInfo": {
        "type": "object",
        "description": "Response wrapper for list endpoints.\nAll list endpoints return responses wrapped in a `data` field.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Key-value entry info (key and timestamps, no value)",
              "required": [
                "key",
                "value",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "description": "When the key was created"
                },
                "key": {
                  "type": "string",
                  "description": "The key name"
                },
                "updated_at": {
                  "type": "string",
                  "description": "When the key was last updated"
                },
                "value": {
                  "type": "string",
                  "description": "The stored value"
                }
              }
            },
            "description": "Array of items returned by the list operation."
          }
        }
      },
...
      "ListResponse_OrganizationResponse": {
        "type": "object",
        "description": "Response wrapper for list endpoints.\nAll list endpoints return responses wrapped in a `data` field.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Response for organization operations",
              "required": [
                "id",
                "name",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When the organization was created"
                },
                "id": {
                  "type": "string",
                  "description": "External identifier (org_<32-hex-chars>)"
                },
                "name": {
                  "type": "string",
                  "description": "Display name"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When the organization was last updated"
                }
              }
            },
            "description": "Array of items returned by the list operation."
          }
        }
      },
      "ListResponse_SecretInfo": {
        "type": "object",
        "description": "Response wrapper for list endpoints.\nAll list endpoints return responses wrapped in a `data` field.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Secret entry info (name and timestamps only, no value)",
              "required": [
                "name",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "description": "When the secret was created"
                },
                "name": {
                  "type": "string",
                  "description": "The secret name"
                },
                "updated_at": {
                  "type": "string",
                  "description": "When the secret was last updated"
                }
              }
            },
            "description": "Array of items returned by the list operation."
          }
        }
      },
...
      "OrganizationResponse": {
        "type": "object",
        "description": "Response for organization operations",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the organization was created"
          },
          "id": {
            "type": "string",
            "description": "External identifier (org_<32-hex-chars>)"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the organization was last updated"
          }
        }
      },
...
      "UpdateOrganizationRequest": {
        "type": "object",
        "description": "Request to update an organization",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The display name of the organization.",
            "example": "Acme Corporation"
          }
        }
      },
...
    }
  },
...
}